+++ applications/bot/src/commands/index.ts
@@ -5,6 +5,7 @@ import { FinanceCommand } from '~/commands/finance/index.js';
 import { InfoCommand } from '~/commands/info/index.js';
 import { MakanickoCommand } from '~/commands/makanicko/index.js';
 import { PollCommand } from '~/commands/poll/index.js';
+import { RefreshEventsCommand } from '~/commands/refreshEvents/index.js';
 import { SummarizeCommand } from '~/commands/summarize/index.js';
 import { SummonCommand } from '~/commands/summon/index.js';
 import { TrainingCommand } from '~/commands/training/index.js';
@@ -18,4 +19,5 @@ export const commandBuilder = Ix.builder
   .add(SummarizeCommand)
   .add(CarpoolCommand)
   .add(TrainingCommand)
-  .add(PollCommand);
+  .add(PollCommand)
+  .add(RefreshEventsCommand);